Split broker message utils in modules
