+++ nucliadb/nucliadb/tests/utils/broker_messages/helpers.py
@@ -1,0 +1,33 @@
+# Copyright (C) 2021 Bosutech XXI S.L.
+#
+# nucliadb is offered under the AGPL v3.0 and as commercial software.
+# For commercial licensing, contact us at info@nuclia.com.
+#
+# AGPL:
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
+#
+
+from nucliadb_protos.resources_pb2 import Classification
+
+
+def labels_to_classifications(labelset: str, labels: list[str]) -> list[Classification]:
+    classifications = [
+        Classification(
+            labelset=labelset,
+            label=label,
+            cancelled_by_user=False,
+        )
+        for label in labels
+    ]
+    return classifications
